requirements.txt
astroid==2.5.2
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
Flask==1.1.2
Flask-PyMongo==2.3.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.4
pymongo==3.11.3
toml==0.10.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1